InTrade 2.0.0
- Currency amounts are BigDecimal, TradeOffer uses BigDecimal as well (API change)
- Player file and escrow are saved at most four times a second instead of on every click
- /trade reload reads files off the main thread
- Database callbacks go through a main thread queue that is also drained on shutdown
- Trade counts are read after pending writes, placeholders no longer lag behind
- Sounds can be changed or turned off in config.yml
- An empty amount clears the currency, as the input hint says
- Javadoc for the API, shorter README

+++ plugin/src/main/java/me/spolzer/intrade/menu/HistoryDetailMenu.java
@@ -1,5 +1,6 @@
 package me.spolzer.intrade.menu;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -59,10 +60,10 @@ public final class HistoryDetailMenu implements InventoryHolder {
         for (int i = 0; i < items.size() && i < 16; i++) inventory.setItem(i / 4 * 9 + column + i % 4, items.get(i));
     }
 
-    private void currencies(Map<String, Double> amounts, int[] slots) {
+    private void currencies(Map<String, BigDecimal> amounts, int[] slots) {
         Messages messages = plugin.messages();
         int i = 0;
-        for (Map.Entry<String, Double> entry : amounts.entrySet()) {
+        for (Map.Entry<String, BigDecimal> entry : amounts.entrySet()) {
             if (i >= slots.length) break;
             Currency currency = plugin.currencies().byId(entry.getKey());
             inventory.setItem(slots[i++], Icons.icon(currency != null ? currency.icon() : Material.PAPER,
